Big refactor and fix.
This change introduces x/crypto/nacl/box instead of RSA which is faster, easier
to use, and probably even more secure.

Also it contains some major refactoring of the code base and especially
the CLI. I really didn't like the global variables that were suggested
in the cobra docs so I got rid of them.

Also this fixes #1 where the public key of a recipient would not be
encrypted in the payload. Originally this was done because of limited
message size in RSA. Luckily, with NaCl/box, this is no longer an issue.